fix(context): let a descriptor keep its own maxCallsPerSession
`buildPullToolDescriptors` documented "overrides maxCallsPerSession from the
pullConfig if it differs from the descriptor default" but implemented
`pullConfig.maxCallsPerSession ?? d.maxCallsPerSession`. The config schema
defaults that key, so the left operand is never undefined and the config value
always won — a descriptor could never carry a per-tool ceiling of its own.

No behaviour changes today (all three registered descriptors use the shared
default), but the next descriptor that sets its own ceiling would have had it
silently discarded.

Treating "equal to the shared default" as "operator configured nothing"
restores the documented precedence. It is ambiguous only for an operator who
explicitly sets the default value, which by definition changes nothing.
DEFAULT_MAX_CALLS_PER_SESSION is exported for that comparison.

The pull-tool describe blocks move to orchestrator-pull-tools.test.ts —
the new test pushed orchestrator.test.ts to 820 lines, over the 800 limit.
