+++ test/unit/context/engine/orchestrator.test.ts
@@ -1,6 +1,5 @@
 import { describe, test, expect, beforeEach } from "bun:test";
 import { ContextOrchestrator, _orchestratorDeps } from "../../../../src/context/engine/orchestrator";
-import { QUERY_NEIGHBOR_DESCRIPTOR, QUERY_FEATURE_CONTEXT_DESCRIPTOR } from "../../../../src/context/engine/pull-tools";
 import { NeutralityLintError } from "@/context";
 import type { ContextRequest, IContextProvider, ContextProviderResult } from "../../../../src/context/engine/types";
 
@@ -330,143 +329,6 @@ describe("ContextOrchestrator.rebuildForAgent()", () => {
     const rebuilt = orch.rebuildForAgent(original, { priorStageDigest: "Updated prior digest." });
     expect(rebuilt.pushMarkdown).toContain("Updated prior digest.");
     expect(rebuilt.manifest.requestId).not.toBe(original.manifest.requestId);
-  });
-});
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Phase 4: pull tools
-// ─────────────────────────────────────────────────────────────────────────────
-
-describe("Phase 4: pull tools", () => {
-  const TDD_IMPLEMENTER_REQUEST: ContextRequest = {
-    storyId: "US-001",
-    repoRoot: "/project",
-    packageDir: "/project",
-    stage: "tdd-implementer",
-    role: "implementer",
-    budgetTokens: 8_000,
-    providerIds: [],
-  };
-
-  test.each([
-    ["pullConfig is absent", undefined],
-    ["pullConfig.enabled is false", { enabled: false, allowedTools: [] as string[], maxCallsPerSession: 5 }],
-  ])("pullTools is empty when %s", async (_label, pullConfig) => {
-    const orch = new ContextOrchestrator([]);
-    const bundle = await orch.assemble({ ...TDD_IMPLEMENTER_REQUEST, pullConfig });
-    expect(bundle.pullTools).toEqual([]);
-  });
-
-
-  test("pullTools items are ToolDescriptor objects; maxCallsPerSession reflects pullConfig override", async () => {
-    const orch = new ContextOrchestrator([]);
-    const bundle = await orch.assemble({
-      ...TDD_IMPLEMENTER_REQUEST,
-      pullConfig: { enabled: true, allowedTools: [], maxCallsPerSession: 3 },
-    });
-    const tool = bundle.pullTools[0]!;
-    expect(typeof tool.name).toBe("string");
-    expect(typeof tool.description).toBe("string");
-    expect(typeof tool.inputSchema).toBe("object");
-    expect(typeof tool.maxCallsPerSession).toBe("number");
-    expect(typeof tool.maxTokensPerCall).toBe("number");
-    expect(tool.maxCallsPerSession).toBe(3);
-  });
-
-  test("allowedTools filter restricts pull tools", async () => {
-    const orch = new ContextOrchestrator([]);
-    const bundle = await orch.assemble({
-      ...TDD_IMPLEMENTER_REQUEST,
-      pullConfig: { enabled: true, allowedTools: ["other_tool"], maxCallsPerSession: 5 },
-    });
-    // query_neighbor is not in allowedTools — filtered out
-    expect(bundle.pullTools).toEqual([]);
-  });
-
-  test("empty allowedTools means all stage-configured tools are allowed; tdd-implementer has query_neighbor", async () => {
-    const orch = new ContextOrchestrator([]);
-    const bundle = await orch.assemble({
-      ...TDD_IMPLEMENTER_REQUEST,
-      pullConfig: { enabled: true, allowedTools: [], maxCallsPerSession: 5 },
-    });
-    expect(bundle.pullTools.length).toBeGreaterThan(0);
-    expect(bundle.pullTools[0]?.name).toBe("query_neighbor");
-  });
-
-  test("stage with no pullToolNames returns empty pullTools even when enabled", async () => {
-    const orch = new ContextOrchestrator([]);
-    const verifyRequest: ContextRequest = {
-      ...TDD_IMPLEMENTER_REQUEST,
-      stage: "verify",
-      pullConfig: { enabled: true, allowedTools: [], maxCallsPerSession: 5 },
-    };
-    const bundle = await orch.assemble(verifyRequest);
-    expect(bundle.pullTools).toEqual([]);
-  });
-
-  test("rebuildForAgent preserves pullTools from original bundle", async () => {
-    const orch = new ContextOrchestrator([]);
-    const original = await orch.assemble({
-      ...TDD_IMPLEMENTER_REQUEST,
-      pullConfig: { enabled: true, allowedTools: [], maxCallsPerSession: 5 },
-    });
-    expect(original.pullTools).toHaveLength(1);
-
-    const rebuilt = orch.rebuildForAgent(original);
-    expect(rebuilt.pullTools).toEqual(original.pullTools);
-    expect(rebuilt.pullTools[0]?.name).toBe(QUERY_NEIGHBOR_DESCRIPTOR.name);
-  });
-});
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Phase 5: review stage pull tools
-// ─────────────────────────────────────────────────────────────────────────────
-
-describe("Phase 5: review stage pull tools", () => {
-  const REVIEW_REQUEST: ContextRequest = {
-    storyId: "US-001",
-    repoRoot: "/project",
-    packageDir: "/project",
-    stage: "review-semantic",
-    role: "reviewer",
-    budgetTokens: 6_000,
-    providerIds: [],
-  };
-
-  test.each(["review-semantic", "review-adversarial"] as const)("%s with pullConfig enabled returns query_feature_context", async (stage) => {
-    const orch = new ContextOrchestrator([]);
-    const bundle = await orch.assemble({
-      ...REVIEW_REQUEST,
-      stage,
-      pullConfig: { enabled: true, allowedTools: [], maxCallsPerSession: 5 },
-    });
-    expect(bundle.pullTools).toHaveLength(1);
-    expect(bundle.pullTools[0]?.name).toBe(QUERY_FEATURE_CONTEXT_DESCRIPTOR.name);
-  });
-
-  test("review-semantic pullConfig disabled returns empty pull tools", async () => {
-    const orch = new ContextOrchestrator([]);
-    const bundle = await orch.assemble({
-      ...REVIEW_REQUEST,
-      pullConfig: { enabled: false, allowedTools: [], maxCallsPerSession: 5 },
-    });
-    expect(bundle.pullTools).toEqual([]);
-  });
-
-  test("pull tool names do not bleed across stages: tdd-implementer lacks query_feature_context, review-semantic lacks query_neighbor", async () => {
-    const orchA = new ContextOrchestrator([]);
-    const bundleA = await orchA.assemble({
-      storyId: "US-001", repoRoot: "/project", packageDir: "/project",
-      stage: "tdd-implementer", role: "implementer", budgetTokens: 8_000, providerIds: [],
-      pullConfig: { enabled: true, allowedTools: [], maxCallsPerSession: 5 },
-    });
-    expect(bundleA.pullTools.map((t) => t.name)).not.toContain("query_feature_context");
-    const orchB = new ContextOrchestrator([]);
-    const bundleB = await orchB.assemble({
-      ...REVIEW_REQUEST,
-      pullConfig: { enabled: true, allowedTools: [], maxCallsPerSession: 5 },
-    });
-    expect(bundleB.pullTools.map((t) => t.name)).not.toContain("query_neighbor");
   });
 });
